Add request creator object to pull creating a request away from the HTTP stuff
